requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
requests==2.31.0
cryptography==41.0.7 
fastapi==0.104.1
uvicorn
mysql-connector-python==8.2.0
